% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{addLabel}
\alias{addLabel}
\title{Labels and captions using \code{bookdown} cross-referencing format}
\usage{
addLabel(caption = "", tag = "tab")
}
\arguments{
\item{caption}{a string with the desired caption.}

\item{tag}{the cross-reference tag used by \code{bookdown}. Recognised
tagas are \code{"tab"}, \code{"fig"} and \code{"eq"}. See
https://bookdown.org/yihui/bookdown}
}
\value{
the full caption with its label in a HTML or LaTeX compatible
format
}
\description{
Makes captions for tables and figures using the \code{bookdown}
cross-referencing format. Meant to be used inside an Rmarkdown chunk.
}
