% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{panble}
\alias{panble}
\title{Pander and kable wrapper function}
\usage{
panble(
  tab,
  caption = "",
  landscape = FALSE,
  panderArgs = list(),
  kableArgs = list(),
  kable_stylingArgs = list(),
  column_specArgs = list()
)
}
\arguments{
\item{tab}{table object compatible with \code{pander::pander} AND
\code{knitr::kable}}

\item{caption}{a caption. Make sure special LaTeX characters are escaped.
\code{bookdown} text references can be useful when formatting text and using
special characters (https://bookdown.org/yihui/bookdown/markdown-extensions-by-bookdown.html#text-references).}

\item{landscape}{if TRUE panderOptions are changed so that the page can be
put in landscape position. Note that this requires adding \verb{\\\\newpage} and
\code{\\begin{landscape}} before the chunk and \code{\\end{landscape}} after the chunk.}

\item{panderArgs}{named list of additional arguments passed to \code{pander}.
Do NOT pass the caption and input table arguments.}

\item{kableArgs}{named list of additional arguments passed to \code{kable}.
Do NOT pass the caption and input table arguments.}

\item{kable_stylingArgs}{named list of additional arguments passed to
\code{kable_styling}. Do NOT pass the input table argument.}

\item{column_specArgs}{named list of additional arguments passed to
\code{column_spec}. Do NOT pass the input table argument.}
}
\value{
a markdown table.
}
\description{
This function overcomes/works around a recent issue with
\code{kable(..., longtable = TRUE)} and \code{kable_styling(..., full_width = TRUE)}
which fail when knitting to PDF (see https://stackoverflow.com/questions/71651334/longtable-t-messes-up-with-scaled-down-table-in-r-markdown-pdf/72511243#72511243).
It uses \code{pander::pander.table} to automatically deal with long/wide tables when knitting to PDF
and \code{knitr::kable} when knitting to HTML.
Note that the chunk option \code{results} must be set to \code{"asis"}.
}
