% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR-related.R
\name{prepLandRRmds}
\alias{prepLandRRmds}
\title{Prepare module .Rmd to render book}
\usage{
prepLandRRmds(modulePath, rebuildCache = FALSE)
}
\arguments{
\item{modulePath}{modules' folder directory}
}
\value{
directories for modified module .Rmd files
}
\description{
Creates modified versions of the modules' .Rmd files,
with YAML headers removed and adapted knitr setup
chunks
}
