% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR-related.R
\name{prepManualRmds}
\alias{prepManualRmds}
\title{Prepare module .Rmd to render book}
\usage{
prepManualRmds(modulePath, rebuildCache = FALSE)
}
\arguments{
\item{modulePath}{modules' folder directory}

\item{rebuildCache}{should cached chunks be re-executed?}
}
\value{
directories for modified module .Rmd files
}
\description{
Creates modified versions of the modules' .Rmd files,
with YAML headers removed and adapted knitr setup
chunks
}
