% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR-related.R
\name{prepManualRmds}
\alias{prepManualRmds}
\title{Prepare module .Rmd to render book}
\usage{
prepManualRmds(modulePath, rebuildCache = FALSE)
}
\arguments{
\item{modulePath}{modules' folder directory. It must be
exactly as in the \verb{_bookdown.yml} lines specifying where
the module .Rmds live. For instance, if in \verb{_bookdown.yml}
the list of module .Rmds is provided as \code{-  modules/XXX.Rmd},
\verb{-  ~/modules/XXX.Rmd}, or \code{-  ../X/modules/XXX.Rmd}
\code{modulePath} must be either \code{modules}, \verb{~/modules} and
\code{../X/modules} respectively. A \code{/} may be appended at the
of \code{modulePath} (e.g. \verb{~/modules/}).
Note that all modules must be in the same directory.}

\item{rebuildCache}{should cached chunks be re-executed?}
}
\value{
file paths of the modified module .Rmd files
}
\description{
Creates modified versions of the modules' .Rmd files,
with YAML headers removed and adapted knitr setup
chunks
}
